requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
tabula-py==2.3.0
PyPDF2==1.26.0
pandas==1.3.0
openpyxl==3.0.7
pdfminer.six==20201018
